Refactor: extract sync time logic to sync_time_index.py module
