+++ backend/routes/video_update_route.py
@@ -1,9 +1,6 @@
 from flask import request, jsonify
-from services.firestore.batch_writer import (
-    get_last_video_sync_time,
-    write_batches_to_firestore,
-    update_last_sync_time,
-)
+from services.firestore.batch_writer import write_batches_to_firestore
+from services.firestore.sync_time_index import get_last_video_sync_time, update_last_sync_time
 from services.youtube.fetcher import get_video_data
 from datetime import datetime, timezone, timedelta
 import logging
